requirements.txt
botbuilder-core
aiohttp
azure-ai-language-conversations
azure-cosmos
azure-core